Project Refactor
Have begun refactoring the project to set the foundation for it to be able to use a new Fraction class, as well as continue to be able to use previously used int, and maybe new types in the future. Currently getting an error on line 67, 97

+++ errorqueue.cpp
@@ -21,7 +21,7 @@ void ErrorQueue::print_errors() {
 };
 
 void ErrorQueue::clear() {
-  for (int i=0; i<errors.size(); i++) {
+  for (double i=0; i<errors.size(); i++) {
     errors.pop();
   }
 }
