+++ lexer.cpp
@@ -9,7 +9,7 @@ Lexer::Lexer(string &input) : input(input), curr(0), next(0) {
 vector<Token> Lexer::tokenize() {  
   vector<Token> tokens;
 
-  while (curr < int(input.length())) {   
+  while (curr < double(input.length())) {   
     Token t;
     skipwhitespace();  
 
@@ -83,7 +83,7 @@ vector<Token> Lexer::tokenize() {
 }
 
 void Lexer::readChar() {
-  if (next >= int(input.length())){
+  if (next >= double(input.length())){
     ch = 0;
   } else {
     ch = input[next];
@@ -93,7 +93,7 @@ void Lexer::readChar() {
 }
 
 char Lexer::peekChar() {
-  if (next >= int(input.length())) {
+  if (next >= double(input.length())) {
     return 0;
   } else {
     return input[next];
@@ -117,8 +117,8 @@ void Lexer::insertInput(string& input) {
 
 // RUNS OVER INPUT LENGTH, has to do with needing space at the end
 string Lexer::readNumber() {
-  int pos = curr;
-  while (curr < int(input.length()) && isDigit(input.at(curr))) {
+  double pos = curr;
+  while (curr < double(input.length()) && isDigit(input.at(curr))) {
     readChar();
   }
 
@@ -127,8 +127,8 @@ string Lexer::readNumber() {
 
 string Lexer::readWord(){
   string s;
-  int pos = curr;
-  while (curr < int(input.length()) && isLetter(input.at(curr))) {
+  double pos = curr;
+  while (curr < double(input.length()) && isLetter(input.at(curr))) {
     readChar();
   }
   return input.substr(pos, curr - pos);
